requirements.txt
Django==2.0
django-ckeditor==5.6.1
django-js-asset==1.1.0
mysqlclient==1.3.14
Pillow==5.4.0
pip==18.1
pytz==2018.7
setuptools==40.6.3
wheel==0.32.3
